Added delete user function, commented some code.
No major changes, still need to cover issues posted on github.
Note to self, add create database and destroy database functions
should be easy to knock off. I'm not sure if it's warranted to
add a create tables function as well, because without a schema...
still maybe I can get a generic schema and create that.

It would probably be more advantageous to finish writing a grant
and revoke permissions function based on a yaml file.